fix conversion warning on MSVC
"src\enc\frame.c(88) : warning C4244: '=' : conversion from 'const double' to 'float', possible loss of data"

+++ src/enc/frame.c
@@ -85,7 +85,7 @@ static float ComputeNextQ(PassStats* const s) {
     s->is_first = 0;
   } else if (s->value != s->last_value) {
     const double slope = (s->target - s->value) / (s->last_value - s->value);
-    dq = slope * (s->last_q - s->q);
+    dq = (float)(slope * (s->last_q - s->q));
   } else {
     dq = 0.;  // we're done?!
   }
